+++ src/controllers/users/deleteUser.js
@@ -1,8 +1,13 @@
 import { PrismaClient } from "@prisma/client";
 import NotFoundError from "../../errors/NotFoundError.js";
+import BadRequestError from "../../errors/BadRequestError.js";
 
 const deleteUser = async (id) => {
   const prisma = new PrismaClient();
+
+  if (!id) {
+    throw new BadRequestError("id is required");
+  }
 
   const deleteUser = await prisma.user.deleteMany({
     where: {
